* Makefile.am: Distribute files ChangeLog, INSTALL-mac.txt,
* coders/svg.c (SVGStartElement): Applied fix from ImageMagick to

* magick/image.c (CloneImage): Applied fix from ImageMagick which

* magick/gem.c (TransformHWB): Replace implementation with

* coders/msl.c (MSLStartElement): Applied fix for missing break

+++ magick/image.c
@@ -1129,8 +1129,8 @@ MagickExport Image *CloneImage(const Image *image,const unsigned long columns,
     }
   clone_image->page.width=columns;
   clone_image->page.height=rows;
-  clone_image->page.x=(long) (columns*image->page.x/clone_image->columns);
-  clone_image->page.y=(long) (rows*image->page.y/clone_image->rows);
+  clone_image->page.x=(long) columns*image->page.x/(long) clone_image->columns;
+  clone_image->page.y=(long) rows*image->page.y/(long) clone_image->rows;
   clone_image->columns=columns;
   clone_image->rows=rows;
   GetCacheInfo(&clone_image->cache);
